backend/requirements.txt
fastapi
uvicorn
openai
openai-guardrails
python-dotenv
httpx
Jinja2
PyYAML
pytest
pytest-cov
